added `nn.Embedding` class & built the backward in autograd, changed a few other files
