+++ axgrad/autograd/_backward.py
@@ -8,10 +8,11 @@
 
 from typing import Literal, Callable
 from ..helpers.functionals import *
-from .functions.binary_ops import __ADD__, __MUL__, __MATMUL__, __POW__, __STACK__, __CONCAT__, __CONV2D__
+from .functions.binary_ops import __ADD__, __MUL__, __MATMUL__, __POW__, __CONV2D__
 from .functions.shape_ops import __TRANSPOSE__, __SWAPAXES__, __RESHAPE__, __SQUEEZE__, __UNSQUEEZE__, __FLATTEN__, __VIEW__, __BROADCAST__
 from .functions.uniary_ops import __SUM__, __MEAN__, __VAR__, __STD__, __EXP__, __RSQRT__, __SQRT__, __LOG__, __CLIP__, __ABS__
 from .functions.activations import __GELU__, __RELU__, __SIGMOID__, __SILU__, __TANH__, __LRELU__
+from .functions.other_ops import __EMBEDD__, __STACK__, __CONCAT__
 from ..helpers.shape import *
 
 class Backward:
@@ -35,14 +36,6 @@ class Backward:
   def conv2d_backwards(first:Literal["tensor"], kernel:Union[Literal["tensor"], list], out:Literal["tensor"], stride:int, padding:int) -> Callable:
     _back = __CONV2D__(first, kernel, out, stride, padding)
     return _back
-  
-  def stack_backwards(out:Literal["tensor"], tensors:Literal["tensor"], axis:int) -> Callable:
-   _back = __STACK__(out, tensors, axis)
-   return _back
-  
-  def concat_backwards(out:Literal["tensor"], tensors:Literal["tensor"], axis:int) -> Callable:
-   _back = __CONCAT__(out, tensors, axis)
-   return _back
 
   ## unary ops backwards:
   def sum_backwards(out: Literal["tensor"], first: Literal["tensor"], axis: Optional[int], keepdims: bool) -> Callable:
@@ -80,6 +73,20 @@ class Backward:
   def sqrt_backwards(out:Literal["tensor"], first:Literal["tensor"]) -> Callable:
     _back = __SQRT__(first, out)
     return _back
+  
+  ## other ops backwards:
+  def embed_backwards(first:Literal["tensor"], indices:list, out:Literal["tensor"]) -> Callable:
+    _back = __EMBEDD__(first, indices, out)
+    return _back
+  
+  def stack_backwards(out:Literal["tensor"], tensors:Literal["tensor"], axis:int) -> Callable:
+   _back = __STACK__(out, tensors, axis)
+   return _back
+  
+  def concat_backwards(out:Literal["tensor"], tensors:Literal["tensor"], axis:int) -> Callable:
+   _back = __CONCAT__(out, tensors, axis)
+   return _back
+
 
   ## shape ops backwards:
   def transpose_backwards(first:Literal["tensor"], out:Literal["tensor"]) -> Callable:
