+++ axgrad/nn/__init__.py
@@ -1,9 +1,8 @@
-from ._linear import Linear
 from ._module import Module
 from ._parameters import Parameter
 from ._activations import *
 from ._functional import functional
 from ._optim import SGD, LARS
-from ._conv import Conv2d
 from ._loss import MSE, MAE
 from ._norm import LayerNorm, BatchNorm
+from ._functions import Linear, Embedding, Conv2d
